feat(wasm): add web target with Room 3.0 OPFS and Firebase JS SDK
Adds a fully working wasmJs (Compose for Web) target across all 22 modules:

- Room 2.8.4 → 3.0.0-alpha04 migration (androidx.room → androidx.room3),
  unified across Android+iOS+wasmJs. nonWasmMain source set removed.
- WebWorkerSQLiteDriver via OPFS for browser persistence (sqlite-wasm-worker
  bundled as local npm package).
- Custom DataStore<Preferences> impl backed by browser localStorage with JSON
  serialization — replaces the broken WebLocalStorage stub.
- Firebase JS SDK facade via init.js (Remote Config + Functions) loaded as
  ESM module; build-time templating from local.properties via the new
  generateWasmInitJs Gradle task.
- Kotlin 2.3.20 + kotlinx-datetime 0.7.1 (resolves IrTypeAlias linker error
  on wasmJs).
- COOP/COEP/CORP webpack-dev-server headers required for SharedArrayBuffer
  and Skiko Web Worker.
- All 21 expect declarations get wasmJs actual stubs (graceful no-op for
  Glance widget, AlarmManager, RevenueCat; real impl for Time, Logger,
  DataStore, DatabaseBuilder, DynamicColor, RemoteConfig).
- CORS headers added to firebase-functions/main.py for register_user and
  restore_credits_after_purchase (deploy still required server-side).
- Playwright e2e harness in e2e/ — 9 tests pass covering bridges, OPFS,
  SAB, DataStore persistence, init.js content; 5 canvas-render tests
  skipped (Playwright wasm streaming limitation, real Chrome unaffected).
- docs/plans/2026-05-07-feat-finish-wasmjs-web-plan.md captures remaining
  Phase 4 (Gemini + Sharing) and Phase 8 (doc-writer) work for next session.

Android assembleDebug PASS — production 1.14.3 not regressed.

+++ composeApp/src/commonMain/kotlin/com/antonchuraev/homesearchchecklist/csat/CsatManager.kt
@@ -5,7 +5,7 @@ import com.antonchuraev.homesearchchecklist.core.datastore.api.AppDatastore
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
-import kotlinx.datetime.Clock
+import kotlin.time.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.todayIn
 
@@ -44,7 +44,7 @@ class CsatManager(private val datastore: AppDatastore) {
     )
 
     private fun todayEpochDays(): Int =
-        Clock.System.todayIn(TimeZone.currentSystemDefault()).toEpochDays()
+        Clock.System.todayIn(TimeZone.currentSystemDefault()).toEpochDays().toInt()
 
     suspend fun onUserAction() {
         val count = datastore.observeInt(KEY_ACTION_COUNT, 0).first()
